Refactoring: moved TTscan to a separate file
